Extract WY from a date

Determine the correct water year based on a calendar date.

calcWaterYear(dateVec)

Arguments

dateVec
vector of dates as character ("YYYY-DD-MM"), Date, or POSIXct. Numeric does not work.

Value

numeric vector indicating the water year

Details

This function calculates a water year based on the USGS definition that a water year starts on October 1 of the year before, and ends on September 30. For example, water year 2015 started on 2014-10-01 and ended on 2015-09-30. See the USGS definition at https://water.usgs.gov/nwc/explain_data.html.

Examples

x <- seq(as.Date("2010-01-01"), as.Date("2010-12-31"), by="month")
waterYear <- calcWaterYear(x)