Construct WQP url for data retrieval

Construct WQP url for data retrieval. This function gets the data from here: https://www.waterqualitydata.us

Usage

constructWQPURL(siteNumbers, parameterCd, startDate, endDate, zip = TRUE)

Arguments

siteNumbers

string or vector of strings USGS site number. This is usually an 8 digit number

parameterCd

string or vector of USGS parameter code. This is usually an 5 digit number.

startDate

character starting date for data retrieval in the form YYYY-MM-DD. Default is "" which indicates retrieval for the earliest possible record.

endDate

character ending date for data retrieval in the form YYYY-MM-DD. Default is "" which indicates retrieval for the latest possible record.

zip

logical to request data via downloading zip file. Default set to TRUE.

Value

url string

Examples

site_id <- '01594440'
startDate <- '1985-01-01'
endDate <- ''
pCode <- c("00060","00010")
url_wqp <- constructWQPURL(paste("USGS",site_id,sep="-"),
           c('01075','00029','00453'),
           startDate,endDate)
url_wqp
#> [1] "https://www.waterqualitydata.us/data/Result/search?siteid=USGS-01594440&pCode=01075;00029;00453&startDateLo=01-01-1985&mimeType=tsv&zip=yes"
charNames <- c("Temperature",
               "Temperature, sample",
               "Temperature, water",
               "Temperature, water, deg F")
obs_url_orig <- constructWQPURL(siteNumbers = c("IIDFG-41WSSPAHS",
                                                "USGS-02352560"), 
                                parameterCd = charNames,
                                startDate,"")
obs_url_orig
#> [1] "https://www.waterqualitydata.us/data/Result/search?siteid=IIDFG-41WSSPAHS;USGS-02352560&characteristicName=Temperature;Temperature%2C%20sample;Temperature%2C%20water;Temperature%2C%20water%2C%20deg%20F&startDateLo=01-01-1985&mimeType=tsv&zip=yes"