Parse the WaterML2 timeseries portion of a waterML2 file

Returns data frame columns of all information with each time series measurement; Anything defined as a default, is returned as an attribute of that data frame.

importWaterML2(input, asDateTime = FALSE, tz = "UTC")

Arguments

input
XML with only the wml2:MeasurementTimeseries node and children

asDateTime
logical, if TRUE returns date and time as POSIXct, if FALSE, character

tz
character to set timezone attribute of datetime. Default is an empty quote, which converts the datetimes to UTC (properly accounting for daylight savings times based on the data's provided time zone offset). Possible values are "America/New_York","America/Chicago", "America/Denver","America/Los_Angeles", "America/Anchorage","America/Honolulu","America/Jamaica","America/Managua","America/Phoenix", and "America/Metlakatla"

Examples

baseURL <- "https://waterservices.usgs.gov/nwis/dv/?format=waterml,2.0"
URL <- paste(baseURL, "sites=01646500",
     "startDT=2014-09-01",
     "endDT=2014-09-08",
     "statCd=00003",
     "parameterCd=00060",sep="&")

timesereies <- importWaterML2(URL, asDateTime=TRUE, tz="UTC")