Data to convert USGS parameter code to characteristic names

Data pulled from Water Quality Portal on November 25, 2014. The data was pulled from https://www.waterqualitydata.us/public_srsnames?mimeType=json.

Value

pCodeToName data frame with information about USGS parameters and how they relate to characteristic names (useful for WQP requests).

Name	Type	Description

parm_cd	character	5-digit USGS parameter code

description	character	Parameter description

characteristicname	character	Characteristic Name

measureunitcode	character	Parameter units

resultsamplefraction	character	Result sample fraction text

resulttemperaturebasis	character	Temperature basis information

resultstatisticalbasis	character	Statistical basis

resulttimebasis	character	Time basis

resultweightbasis	character	Weight basis

resultparticlesizebasis	character	Particle size basis

last_rev_dt	character	Latest revision of information

Examples

head(pCodeToName[,1:2])

#>   parm_cd                                 description
#> 1   00004                          Stream width, feet
#> 2   00010         Temperature, water, degrees Celsius
#> 3   00011      Temperature, water, degrees Fahrenheit
#> 4   00020           Temperature, air, degrees Celsius
#> 5   00021        Temperature, air, degrees Fahrenheit
#> 6   00025 Barometric pressure, millimeters of mercury