Construct WQP url for data retrieval

Construct WQP url for data retrieval. This function gets the data from here: https://www.waterqualitydata.us

constructWQPURL(siteNumbers, parameterCd, startDate, endDate,
  zip = FALSE)

Arguments

siteNumbers
string or vector of strings USGS site number. This is usually an 8 digit number

parameterCd
string or vector of USGS parameter code. This is usually an 5 digit number.

startDate
character starting date for data retrieval in the form YYYY-MM-DD. Default is "" which indicates retrieval for the earliest possible record.

endDate
character ending date for data retrieval in the form YYYY-MM-DD. Default is "" which indicates retrieval for the latest possible record.

zip
logical to request data via downloading zip file. Default set to FALSE.

Value

url string

Examples

site_id <- '01594440'
startDate <- '1985-01-01'
endDate <- ''
pCode <- c("00060","00010")
url_wqp <- constructWQPURL(paste("USGS",site_id,sep="-"),
           c('01075','00029','00453'),
           startDate,endDate)