Retrieval functions for USGS and EPA data

Package:

dataRetrieval

Type:	Package

License:

Unlimited for this package, dependencies have more restrictive licensing.

Copyright:	This software is in the public domain because it contains materials

that originally came from the United States Geological Survey, an agency of	the United States Department of Interior. For more information, see the

official USGS copyright policy at	https://www.usgs.gov/visual-id/credit_usgs.html

LazyLoad:	yes

Details

Retrieval functions for USGS and EPA hydrologic and water quality data.

Please see https://pubs.er.usgs.gov/publication/tm4A10 for more information.