Retrieve features from the NLDI

Provides a formal query to the Network Linked Data Index. The function is useful for topology and location based feature discovery. A user must supply a starting feature, and can add optional navigation direction(s), and features to identify on the navigated network. Valid starting options can be given by one of the following arguments: comid, nwis, huc12, wqp, location, and start.

findNLDI(
  comid = NULL,
  nwis = NULL,
  wqp = NULL,
  huc12 = NULL,
  location = NULL,
  origin = NULL,
  nav = NULL,
  find = NULL,
  distance_km = 100,
  no_sf = FALSE
)

Arguments

comid
an NHDPlusV2 COMID

nwis
a USGS NWIS siteID

wqp
a water quality point ID

huc12
a HUC12 ID

location
Coordinate pair in WGS84 GCS provided as a numeric vector ordered lng/lat

origin
a named list specifying a feature type and ID (e.g. list("comid" = 101))

nav
where to navigate from the starting point ("UM", "UT", DM", "DD")

find
what resources to find along the navigation path(s) (see get_nldi_sources()$source). Can also include 'basin', which will return the upstream basin of the starting feature

distance_km
how far to look along the navigation path in kilometers (default = 100)

no_sf
if available, should `sf` be used for parsing, defaults to `TRUE` if `sf` is locally installed

Value

a list of data.frames

Examples

# Find Features / Define origin features

## Find feature by COMID
 findNLDI(comid = 101)

#> $origin
#> Simple feature collection with 1 feature and 3 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -94.64845 ymin: 31.0838 xmax: -94.62997 ymax: 31.09915
#> geographic CRS: WGS 84
#> # A tibble: 1 x 4
#>   sourceName   identifier comid                                         geometry
#>   <chr>        <chr>      <chr>                                 <LINESTRING [°]>
#> 1 NHDPlus com… 101        101   (-94.64845 31.09915, -94.64803 31.09871, -94.64…
#> 

## Find feature by NWIS ID
 findNLDI(nwis = '11120000')

#> $origin
#> Simple feature collection with 1 feature and 5 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -119.8118 ymin: 34.42472 xmax: -119.8118 ymax: 34.42472
#> geographic CRS: WGS 84
#> # A tibble: 1 x 6
#>   sourceName identifier comid        X     Y             geometry
#> * <chr>      <chr>      <chr>    <dbl> <dbl>          <POINT [°]>
#> 1 NWIS Sites 11120000   17595429 -120.  34.4 (-119.8118 34.42472)
#> 

## Find feature by WQP ID
 findNLDI(wqp = 'USGS-04024315')

#> $origin
#> Simple feature collection with 1 feature and 5 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -92.23492 ymin: 46.50355 xmax: -92.23492 ymax: 46.50355
#> geographic CRS: WGS 84
#> # A tibble: 1 x 6
#>   sourceName           identifier    comid       X     Y             geometry
#>   <chr>                <chr>         <chr>   <dbl> <dbl>          <POINT [°]>
#> 1 Water Quality Portal USGS-04024315 1800305 -92.2  46.5 (-92.23492 46.50355)
#> 

## Find feature by LOCATION
 findNLDI(location = c(-115,40))

#> $origin
#> Simple feature collection with 1 feature and 3 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -115.0326 ymin: 40.04013 xmax: -115.0182 ymax: 40.05183
#> geographic CRS: WGS 84
#> # A tibble: 1 x 4
#>   sourceName   identifier comid                                         geometry
#>   <chr>        <chr>      <chr>                                 <LINESTRING [°]>
#> 1 NHDPlus com… 946060315  94606… (-115.0182 40.05183, -115.0185 40.05176, -115.…
#> 

## GENERAL ORIGIN: COMID
 findNLDI(origin = list("comid" = 101))

#> $origin
#> Simple feature collection with 1 feature and 3 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -94.64845 ymin: 31.0838 xmax: -94.62997 ymax: 31.09915
#> geographic CRS: WGS 84
#> # A tibble: 1 x 4
#>   sourceName   identifier comid                                         geometry
#>   <chr>        <chr>      <chr>                                 <LINESTRING [°]>
#> 1 NHDPlus com… 101        101   (-94.64845 31.09915, -94.64803 31.09871, -94.64…
#> 

## GENERAL ORIGIN: WaDE
 findNLDI(origin = list("wade" = 'CA_45206'))

#> $origin
#> Simple feature collection with 1 feature and 5 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -119.7917 ymin: 34.47273 xmax: -119.7917 ymax: 34.47273
#> geographic CRS: WGS 84
#> # A tibble: 1 x 6
#>   sourceName                identifier comid       X     Y              geometry
#>   <chr>                     <chr>      <chr>   <dbl> <dbl>           <POINT [°]>
#> 1 Water Data Exchange 2.0 … CA_45206   175952… -120.  34.5  (-119.7917 34.47273)
#> 

# Navigation
# UPPER MAINSTEM of USGS-11120000
 str(findNLDI(nwis = '11120000', nav = "UM"), max.level = 1)

#> List of 2
#>  $ origin: tibble [1 × 6] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:5] "sourceName" "identifier" "comid" "X" ...
#>  $ UM    : tibble [4 × 2] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA
#>   .. ..- attr(*, "names")= chr "nhdplus_comid"

# MULTI-REQUEST
# UPPER MAINSTEM and TRIBUTARY of USGS-11120000
 str(findNLDI(nwis = '11120000', nav = c("UT", "UM")), max.level = 1)

#> List of 3
#>  $ origin: tibble [1 × 6] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:5] "sourceName" "identifier" "comid" "X" ...
#>  $ UT    : tibble [25 × 2] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA
#>   .. ..- attr(*, "names")= chr "nhdplus_comid"
#>  $ UM    : tibble [4 × 2] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA
#>   .. ..- attr(*, "names")= chr "nhdplus_comid"

# Discover Features

## Find feature(s) on the upper tributary of USGS-11120000
 str(findNLDI(nwis = '11120000', nav = "UT", find = c("nwis", "wqp")), max.level = 1)

#> List of 4
#>  $ origin     : tibble [1 × 6] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:5] "sourceName" "identifier" "comid" "X" ...
#>  $ UT         : tibble [25 × 2] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA
#>   .. ..- attr(*, "names")= chr "nhdplus_comid"
#>  $ UT_nwissite: tibble [2 × 6] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:5] "sourceName" "identifier" "comid" "X" ...
#>  $ UT_WQP     : tibble [64 × 6] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:5] "sourceName" "identifier" "comid" "X" ...

## Find upstream basin boundary of USGS-11120000
 str(findNLDI(nwis = '11120000',  find = "basin"), max.level = 1)

#> List of 2
#>  $ origin: tibble [1 × 6] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:5] "sourceName" "identifier" "comid" "X" ...
#>  $ basin : tibble [1 × 1] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: 
#>   .. ..- attr(*, "names")= chr(0) 

# Control Distance

## Limit search to 100 km
 str(findNLDI(comid = 101, nav = "DM", find = c("nwis", "wqp"), distance_km = 100), max.level = 1)

#> List of 4
#>  $ origin     : tibble [1 × 4] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA
#>   .. ..- attr(*, "names")= chr [1:3] "sourceName" "identifier" "comid"
#>  $ DM         : tibble [43 × 2] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA
#>   .. ..- attr(*, "names")= chr "nhdplus_comid"
#>  $ DM_nwissite: tibble [1 × 6] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:5] "sourceName" "identifier" "comid" "X" ...
#>  $ DM_WQP     : tibble [5 × 6] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:5] "sourceName" "identifier" "comid" "X" ...

## Convert returned list of data.frames to list of spatial features
# \donttest{
 nldi = findNLDI(nwis = '11120000', nav = "UT", find = c("nwis", "wqp"))
 str(nldi, max.level = 1)

#> List of 4
#>  $ origin     : tibble [1 × 6] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:5] "sourceName" "identifier" "comid" "X" ...
#>  $ UT         : tibble [25 × 2] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA
#>   .. ..- attr(*, "names")= chr "nhdplus_comid"
#>  $ UT_nwissite: tibble [2 × 6] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:5] "sourceName" "identifier" "comid" "X" ...
#>  $ UT_WQP     : tibble [64 × 6] (S3: sf/tbl_df/tbl/data.frame)
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:5] "sourceName" "identifier" "comid" "X" ...

 # }