Retrieve features from the NLDI

Provides a formal query to the Network Linked Data Index. The function is useful for topology and location based feature discovery. A user must supply a starting feature, and can add optional navigation direction(s), and features to identify on the navigated network. Valid starting options can be given by one of the following arguments: comid, nwis, huc12, wqp, location, and start.

findNLDI(
  comid = NULL,
  nwis = NULL,
  wqp = NULL,
  huc12 = NULL,
  location = NULL,
  origin = NULL,
  nav = NULL,
  find = NULL,
  distance_km = 100,
  no_sf = FALSE
)

Arguments

comid
an NHDPlusV2 COMID

nwis
a USGS NWIS siteID

wqp
a water quality point ID

huc12
a HUC12 ID

location
Coordinate pair in WGS84 GCS provided as a numeric vector ordered lng/lat

origin
a named list specifying a feature type and ID (e.g. list("comid" = 101))

nav
where to navigate from the starting point ("UM", "UT", DM", "DD")

find
what resources to find along the navigation path(s) (see get_nldi_sources()$source). Can also include 'basin', which will return the upstream basin of the starting feature

distance_km
how far to look along the navigation path in kilometers (default = 100)

no_sf
if available, should `sf` be used for parsing, defaults to `TRUE` if `sf` is locally installed

Value

a list of data.frames

Examples

# Find Features / Define origin features

## Find feature by COMID
 findNLDI(comid = 101)

#> $origin
#>      sourceName identifier                       geometry
#> 1 NHDPlus comid        101 LINESTRING (-94.64845 31.09...
#> 

## Find feature by NWIS ID
 findNLDI(nwis = '11120000')

#> $origin
#>   sourceName identifier         X        Y                   geometry
#> 1 NWIS Sites   11120000 -119.8118 34.42472 POINT (-119.8118 34.42472)
#> 

## Find feature by WQP ID
 findNLDI(wqp = 'USGS-04024315')

#> $origin
#>             sourceName    identifier         X        Y
#> 1 Water Quality Portal USGS-04024315 -92.23492 46.50355
#>                     geometry
#> 1 POINT (-92.23492 46.50355)
#> 

## Find feature by LOCATION
 findNLDI(location = c(-115,40))

#> $origin
#>      sourceName identifier                       geometry
#> 1 NHDPlus comid  946060315 LINESTRING (-115.0182 40.05...
#> 

## GENERAL ORIGIN: COMID
 findNLDI(origin = list("comid" = 101))

#> $origin
#>      sourceName identifier                       geometry
#> 1 NHDPlus comid        101 LINESTRING (-94.64845 31.09...
#> 

## GENERAL ORIGIN: WaDE
 findNLDI(origin = list("wade" = 'CA_45206'))

#> $origin
#>                      sourceName identifier         X        Y
#> 1 Water Data Exchange 2.0 Sites   CA_45206 -119.7917 34.47273
#>                     geometry
#> 1 POINT (-119.7917 34.47273)
#> 

# Navigation
# UPPER MAINSTEM of USGS-11120000
 str(findNLDI(nwis = '11120000', nav = "UM"), max.level = 1)

#> List of 2
#>  $ origin:'data.frame':	1 obs. of  5 variables:
#>  $ UM    :'data.frame':	4 obs. of  2 variables:

# MULTI-REQUEST
# UPPER MAINSTEM and TRIBUTARY of USGS-11120000
 str(findNLDI(nwis = '11120000', nav = c("UT", "UM")), max.level = 1)

#> List of 3
#>  $ origin:'data.frame':	1 obs. of  5 variables:
#>  $ UT    :'data.frame':	25 obs. of  2 variables:
#>  $ UM    :'data.frame':	4 obs. of  2 variables:

# Discover Features

## Find feature(s) on the upper tributary of USGS-11120000
 str(findNLDI(nwis = '11120000', nav = "UT", find = c("nwis", "wqp")), max.level = 1)

#> List of 4
#>  $ origin     :'data.frame':	1 obs. of  5 variables:
#>  $ UT         :'data.frame':	25 obs. of  2 variables:
#>  $ UT_nwissite:'data.frame':	2 obs. of  5 variables:
#>  $ UT_WQP     :'data.frame':	64 obs. of  5 variables:

## Find upstream basin boundary of USGS-11120000
 str(findNLDI(nwis = '11120000',  find = "basin"), max.level = 1)

#> List of 2
#>  $ origin:'data.frame':	1 obs. of  5 variables:
#>  $ basin :'data.frame':	1 obs. of  1 variable:

# Control Distance

## Limit search to 100 km
 str(findNLDI(comid = 101, nav = "DM", find = c("nwis", "wqp"), distance_km = 100), max.level = 1)

#> List of 4
#>  $ origin     :'data.frame':	1 obs. of  3 variables:
#>  $ DM         :'data.frame':	43 obs. of  2 variables:
#>  $ DM_nwissite:'data.frame':	1 obs. of  5 variables:
#>  $ DM_WQP     :'data.frame':	5 obs. of  5 variables:

## Convert returned list of data.frames to list of spatial features
# \donttest{
 nldi = findNLDI(nwis = '11120000', nav = "UT", find = c("nwis", "wqp"))
 str(lapply(nldi, sf::st_as_sf), max.level = 1)

#> List of 4
#>  $ origin     :Classes ‘sf’ and 'data.frame':	1 obs. of  5 variables:
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:4] "sourceName" "identifier" "X" "Y"
#>  $ UT         :Classes ‘sf’ and 'data.frame':	25 obs. of  2 variables:
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA
#>   .. ..- attr(*, "names")= chr "nhdplus_comid"
#>  $ UT_nwissite:Classes ‘sf’ and 'data.frame':	2 obs. of  5 variables:
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:4] "sourceName" "identifier" "X" "Y"
#>  $ UT_WQP     :Classes ‘sf’ and 'data.frame':	64 obs. of  5 variables:
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:4] "sourceName" "identifier" "X" "Y"

 # }