Function to return data from the National Ground Water Monitoring Network waterML2 format

This function accepts a url parameter for a WaterML2 getObservation. This function is still under development, but the general functionality is correct.

Usage

importNGWMN(input, asDateTime = FALSE, tz = "UTC")

Arguments

input

character or raw, containing the url for the retrieval or a path to the data file, or raw XML.

asDateTime

logical, if TRUE returns date and time as POSIXct, if FALSE, character

tz

character to set timezone attribute of dateTime. Default is "UTC", and converts the date times to UTC, properly accounting for daylight savings times based on the data's provided time zone offset. Possible values to provide are "America/New_York", "America/Chicago", "America/Denver", "America/Los_Angeles", "America/Anchorage", as well as the following which do not use daylight savings time: "America/Honolulu", "America/Jamaica", "America/Managua", "America/Phoenix", and "America/Metlakatla". See also OlsonNames() for more information on time zones.

Value

mergedDF a data frame source, time, value, uom, uomTitle, comment, gmlID

Examples

# \donttest{
obs_url <- paste("https://cida.usgs.gov/ngwmn_cache/sos?request=GetObservation",
  "service=SOS", "version=2.0.0",
  "observedProperty=urn:ogc:def:property:OGC:GroundWaterLevel",
  "responseFormat=text/xml",
  "featureOfInterest=VW_GWDP_GEOSERVER.USGS.403836085374401",
  sep = "&"
)

#data_returned <- importNGWMN(obs_url)
# }