Data to convert USGS parameter code to characteristic names

Data pulled from Water Quality Portal on October 13, 2020. The data was pulled from https://www.waterqualitydata.us/public_srsnames?mimeType=json.

Value

pCodeToName data frame with information about USGS parameters and how they relate to characteristic names (useful for WQP requests).

Name	Type	Description

parm_cd	character	5-digit USGS parameter code

description	character	Parameter description

characteristicname	character	Characteristic Name

measureunitcode	character	Parameter units

resultsamplefraction	character	Result sample fraction text

resulttemperaturebasis	character	Temperature basis information

resultstatisticalbasis	character	Statistical basis

resulttimebasis	character	Time basis

resultweightbasis	character	Weight basis

resultparticlesizebasis	character	Particle size basis

last_rev_dt	character	Latest revision of information

Examples

head(pCodeToName[,1:2])

#>   parm_cd                            description
#> 1       4                     Stream width, feet
#> 2      10    Temperature, water, degrees Celsius
#> 3      11 Temperature, water, degrees Fahrenheit
#> 4      14  Wet bulb temperature, degrees Celsius
#> 5      20      Temperature, air, degrees Celsius
#> 6      21   Temperature, air, degrees Fahrenheit