Daily Value USGS NWIS Data Retrieval

Imports data from NWIS web service. This function gets the data from here: https://waterservices.usgs.gov/

readNWISdv(siteNumbers, parameterCd, startDate = "", endDate = "",
  statCd = "00003")

Arguments

siteNumbers
character USGS site number. This is usually an 8 digit number. Multiple sites can be requested with a character vector.

parameterCd
character of USGS parameter code(s). This is usually an 5 digit number.

startDate
character starting date for data retrieval in the form YYYY-MM-DD. Default is "" which indicates retrieval for the earliest possible record. Date arguments are always specified in local time.

endDate
character ending date for data retrieval in the form YYYY-MM-DD. Default is "" which indicates retrieval for the latest possible record. Date arguments are always specified in local time.

statCd
character USGS statistic code. This is usually 5 digits. Daily mean (00003) is the default.

Value

A data frame with the following columns:

Name	Type

Description		agency

character	The NWIS code for the agency reporting the data

site	character	The USGS site number

Date	Date

The date of the value		code

character	Any codes that qualify the corresponding value

value	numeric	The numeric value for the parameter

Name

Note that code and value are repeated for the parameters requested. The names are of the form X_D_P_S, where X is literal, D is an option description of the parameter, P is the parameter code, and S is the statistic code (if applicable).

There are also several useful attributes attached to the data frame:

Name	Type

Description		url

character	The url used to generate the data

siteInfo	data.frame	A data frame containing information on the requested sites

variableInfo	data.frame

A data frame containing information on the requested parameters		statisticInfo

data.frame	A data frame containing information on the requested statistics on the data

queryTime	POSIXct	The time the data was returned

Name

See also

Examples

site_id <- '04085427'
startDate <- '2012-01-01'
endDate <- '2012-06-30'
pCode <- '00060'
# NOT RUN {
rawDailyQ <- readNWISdv(site_id,pCode, startDate, endDate)
rawDailyQAndTempMeanMax <- readNWISdv(site_id,c('00010','00060'),
       startDate, endDate, statCd=c('00001','00003'))
rawDailyQAndTempMeanMax <- renameNWISColumns(rawDailyQAndTempMeanMax)
rawDailyMultiSites<- readNWISdv(c("01491000","01645000"),c('00010','00060'),
       startDate, endDate, statCd=c('00001','00003'))
# Site with no data:
x <- readNWISdv("10258500","00060", "2014-09-08", "2014-09-14")
names(attributes(x))
attr(x, "siteInfo")
attr(x, "variableInfo")

site <- "05212700"
notActive <- readNWISdv(site, "00060", "2014-01-01","2014-01-07")
# }