USGS Site File Data Retrieval

Imports data from USGS site file site. This function gets data from here: https://waterservices.usgs.gov/

Usage

readNWISsite(siteNumbers)

Arguments

siteNumbers

character USGS site number (or multiple sites). This is usually an 8 digit number

Value

A data frame with at least the following columns:

Name	Type	Description

agency_cd	character	The NWIS code for the agency reporting the data

site_no	character	The USGS site number

station_nm	character	Site name

site_tp_cd	character	Site type

lat_va	numeric	DMS latitude

long_va	numeric	DMS longitude

dec_lat_va	numeric	Decimal latitude

dec_long_va	numeric	Decimal longitude

coord_meth_cd	character	Latitude-longitude method

coord_acy_cd	character	Latitude-longitude accuracy

coord_datum_cd	character	Latitude-longitude datum

dec_coord_datum_cd	character	Decimal Latitude-longitude datum

district_cd	character	District code

state_cd	character	State code

county_cd	character	County code

country_cd	character	Country code

land_net_ds	character	Land net location description

map_nm	character	Name of location map

map_scale_fc	character	Scale of location map

alt_va	numeric	Altitude of Gage/land surface

alt_meth_cd	character	Method altitude determined

alt_acy_va	numeric	Altitude accuracy

alt_datum_cd	character	Altitude datum

huc_cd	character	Hydrologic unit code

basin_cd	character	Drainage basin code

topo_cd	character	Topographic setting code

instruments_cd	character	Flags for instruments at site

construction_dt	character	Date of first construction

inventory_dt	character	Date site established or inventoried

drain_area_va	numeric	Drainage area

contrib_drain_area_va	numeric	Contributing drainage area

tz_cd	character	Time Zone abbreviation

local_time_fg	character	Site honors Daylight Savings Time

reliability_cd	character	Data reliability code

gw_file_cd	character	Data-other GW files

nat_aqfr_cd	character	National aquifer code

aqfr_cd	character	Local aquifer code

aqfr_type_cd	character	Local aquifer type code

well_depth_va	numeric	Well depth

hole_depth_va	numeric	Hole depth

depth_src_cd	character	Source of depth data

project_no	character	Project number

There are also several useful attributes attached to the data frame:

Name	Type	Description

url	character	The url used to generate the data

queryTime	POSIXct	The time the data was returned

comment	character	Header comments from the RDB file

Examples

# \donttest{

siteINFO <- readNWISsite("05114000")
siteINFOMulti <- readNWISsite(c("05114000", "09423350"))
# }