Summary of Data Available from Water Quality Portal

Returns a list of sites with year-by-year information on what data is available. The function gets the data from: https://www.waterqualitydata.us. Arguments to the function should be based on https://www.waterqualitydata.us/webservices_documentation. The information returned from this function describes the available data at the WQP sites, and some metadata on the sites themselves.

Usage

readWQPsummary(...)

Arguments

...

see https://www.waterqualitydata.us/webservices_documentation for a complete list of options. A list of arguments can also be supplied. One way to figure out how to construct a WQP query is to go to the "Advanced" form in the Water Quality Portal: https://www.waterqualitydata.us/#mimeType=csv&providers=NWIS&providers=STEWARDS&providers=STORET Use the form to discover what parameters are available. Once the query is set in the form, scroll down to the "Query URL". You will see the parameters after "https://www.waterqualitydata.us/#". For example, if you chose "Nutrient" in the Characteristic Group dropdown, you will see characteristicType=Nutrient in the Query URL. The corresponding argument for dataRetrieval is characteristicType = "Nutrient". dataRetrieval users do not need to include mimeType, zip, and providers is optional (these arguments are picked automatically).

Value

A data frame with at least the following columns:

Name	Type	Description

"Provider"	character	Providing database.

"MonitoringLocationIdentifier"	character	A designator used to describe the unique name, number, or code assigned to identify the monitoring location.

"YearSummarized"	numeric	The year of the summary

"CharacteristicType"	character	CharacteristicType

"CharacteristicName"	character	The object, property, or substance which is evaluated or enumerated by either a direct field measurement, a direct field observation, or by laboratory analysis of material collected in the field.

"ActivityCount"	numeric	The number of times the location was sampled

"ResultCount"	numeric	The number of individual data results.

"LastResultSubmittedDate"	Date	Date when data was last submitted.

"OrganizationIdentifier"	character	A designator used to uniquely identify a unique business establishment within a context.

"OrganizationFormalName"	character	The legal designator (i.e. formal name) of an organization.

"MonitoringLocationName	character	MonitoringLocationName

"MonitoringLocationTypeName"	character	MonitoringLocationTypeName

"ResolvedMonitoringLocationTypeName"	character

"HUCEightDigitCode"	character	8-digit HUC id.

"MonitoringLocationUrl"	character	URL to monitoring location.

"CountyName"	character	County of sampling location.

"StateName"	character	State of sampling location.

"MonitoringLocationLatitude"	numeric	latitude of sampling location.

"MonitoringLocationLongitude"	numeric	longitude of sampling location.

See also

Examples

# \donttest{
# Summary of a single site for the last 5 years:
site_5 <- readWQPsummary(
  siteid = "USGS-07144100",
  summaryYears = 5
)

# Summary of a single site for the full period of record:
site_all <- readWQPsummary(
  siteid = "USGS-07144100",
  summaryYears = "all"
)

# Summary of the data available from streams in a single county:
dane_county_data <- readWQPsummary(
  countycode = "US:55:025",
  summaryYears = 5,
  siteType = "Stream"
)

# Summary of the data all available from lakes in a single county:
lake_sites <- readWQPsummary(
  siteType = "Lake, Reservoir, Impoundment",
  countycode = "US:55:025"
)

# Summary of the data available for the last 5 years in New Jersey:
state1 <- readWQPsummary(
  statecode = "NJ",
  summaryYears = 5,
  siteType = "Stream"
)
# }