USGS data availability

Imports a table of available parameters, period of record, and count. See https://waterservices.usgs.gov/rest/Site-Service.html for more information.

whatNWISdata(...)

Arguments

...
see https://waterservices.usgs.gov/rest/Site-Service.html for a complete list of options. A list of arguments can also be supplied.

Value

A data frame with the following columns:

Name	Type	Description

agency_cd	character	The NWIS code for the agency reporting the data

site_no	character	The USGS site number

station_nm	character	Site name

site_tp_cd	character	Site type

dec_lat_va	numeric	Decimal latitude

dec_long_va	numeric	Decimal longitude

coord_acy_cd	character	Latitude-longitude accuracy

dec_coord_datum_cd	character	Decimal Latitude-longitude datum

alt_va	character	Altitude of Gage or land surface

alt_acy_va	character	Altitude accuracy

alt_datum_cd	character	Altitude datum

huc_cd	character	Hydrologic unit code

data_type_cd	character	Data type

parm_cd	character	Parameter code

stat_cd	character	Statistical code

dd_nu	character	Internal database key

loc_web_ds	character	Additional measurement description

medium_grp_cd	character	Medium group code

parm_grp_cd	character	Parameter group code

srs_id	character	SRS ID

access_cd	character	Access code

begin_date	Date	Begin date

end_date	Date	End date

count_nu	integer	Record count

parameter_group_nm	character	Parameter group name

parameter_nm	character	Parameter name

casrn	character	Chemical Abstracts Service (CAS) Registry Number

srsname	character	Substance Registry Services

parameter_units	character	Parameter units

There are also several useful attributes attached to the data frame:

Name	Type	Description

url	character	The url used to generate the data

comment	character	Header comments from the RDB file

queryTime	POSIXct	The time the data was returned

Examples

# \donttest{
availableData <- whatNWISdata(siteNumber = '05114000')
# To find just unit value ('instantaneous') data:
uvData <- whatNWISdata(siteNumber = '05114000',service="uv")
uvDataMulti <- whatNWISdata(siteNumber = c('05114000','09423350'),service=c("uv","dv"))
flowAndTemp <- whatNWISdata(stateCd = "WI", service = "uv", 
                             parameterCd = c("00060","00010"),
                             statCd = "00003")
# }