Site Data Import from Water Quality Portal

Returns a list of sites from the Water Quality Portal web service. This function gets the data from: https://www.waterqualitydata.us. Arguments to the function should be based on https://www.waterqualitydata.us/webservices_documentation. The return from this function returns the basic metadata on WQP sites. It is generally faster than the whatWQPdata function, but does not return information on what data was collected at the site.

Usage

whatWQPsamples(..., convertType = TRUE)

whatWQPmetrics(..., convertType = TRUE)

whatWQPsites(...)

Arguments

...

see https://www.waterqualitydata.us/webservices_documentation for a complete list of options. A list of arguments can also be supplied. One way to figure out how to construct a WQP query is to go to the "Advanced" form in the Water Quality Portal: https://www.waterqualitydata.us/#mimeType=csv&providers=NWIS&providers=STEWARDS&providers=STORET Use the form to discover what parameters are available. Once the query is set in the form, scroll down to the "Query URL". You will see the parameters after "https://www.waterqualitydata.us/#". For example, if you chose "Nutrient" in the Characteristic Group dropdown, you will see characteristicType=Nutrient in the Query URL. The corresponding argument for dataRetrieval is characteristicType = "Nutrient". dataRetrieval users do not need to include mimeType, zip, and providers is optional (these arguments are picked automatically).

convertType

logical, defaults to TRUE. If TRUE, the function will convert the data to dates, datetimes, numerics based on a standard algorithm. If false, everything is returned as a character.

Value

A data frame with at least the following columns:

Name	Type	Description

OrganizationIdentifier	character	A designator used to uniquely identify a unique business establishment within a context.

OrganizationFormalName	character	The legal designator (i.e. formal name) of an organization.

MonitoringLocationIdentifier	character	A designator used to describe the unique name, number, or code assigned to identify the monitoring location.

MonitoringLocationName	character	The designator specified by the sampling organization for the site at which sampling or other activities are conducted.

MonitoringLocationTypeName	character	The descriptive name for a type of monitoring location.

MonitoringLocationDescriptionText	character	Text description of the monitoring location.

HUCEightDigitCode	character	The 8 digit federal code used to identify the hydrologic unit of the monitoring location to the cataloging unit level of precision.

DrainageAreaMeasure/MeasureValue *	character	The drainage basin of a lake, stream, wetland, or estuary site. Measure value is given in the units stored in DrainageAreaMeasure/MeasureUnitCode.

DrainageAreaMeasure/MeasureUnitCode *	character	The code that represents the unit for measuring the item.

ContributingDrainageAreaMeasure/MeasureValue *	character	The contributing drainage area of a lake, stream, wetland, or estuary site. Measure value is given in the units stored in ContributingDrainageAreaMeasure/MeasureUnitCode.

ContributingDrainageAreaMeasure/MeasureUnitCode *	character	The code that represents the unit for measuring the item.

LatitudeMeasure	numeric	The measure of the angular distance on a meridian north or south of the equator.

LongitudeMeasure	numeric	The measure of the angular distance on a meridian east or west of the prime meridian.

SourceMapScaleNumeric	character	The number that represents the proportional distance on the ground for one unit of measure on the map or photo.

HorizontalAccuracyMeasure/MeasureValue *	character	The horizontal measure of the relative accuracy of the latitude and longitude coordinates." Measure value is given in the units stored in HorizontalAccuracyMeasure/MeasureUnitCode.

HorizontalAccuracyMeasure/MeasureUnitCode *	character	The code that represents the unit for measuring the item.

HorizontalCollectionMethodName	character	The name that identifies the method used to determine the latitude and longitude coordinates for a point on the earth.

HorizontalCoordinateReferenceSystemDatumName	character	The name that describes the reference datum used in determining latitude and longitude coordinates.

VerticalMeasure/MeasureValue	character	The measure of elevation (i.e., the altitude), above or below a reference datum. Measure value is given in the units stored in VerticalMeasure/MeasureUnitCode.

VerticalMeasure/MeasureUnitCode	character	The code that represents the unit for measuring the item.

VerticalAccuracyMeasure/MeasureValue *	character	The vertical measure of the relative accuracy of the latitude and longitude coordinates. Measure value is given in the units stored in VerticalAccuracyMeasure/MeasureUnitCode.

VerticalAccuracyMeasure/MeasureUnitCode *	character	The code that represents the unit for measuring the item.

VerticalCollectionMethodName	character	The name that identifies the method used to collect the vertical measure (i.e. the altitude) of a reference point.

VerticalCoordinateReferenceSystemDatumName	character	The name of the reference datum used to determine the vertical measure (i.e., the altitude).

CountryCode	character	A code designator used to identify a primary geopolitical unit of the world.

StateCode	character	A code designator used to identify a principal administrative subdivision of the United States, Canada, or Mexico.

CountyCode	character	A code designator used to identify a U.S. county or county equivalent.

AquiferName *	character	Name of the aquifer in which the well is completed.

FormationTypeText *	character	Name of the primary formation or soils unit, in which the well is completed.

AquiferTypeName *	character	The type of aquifer, such as confined or unconfined.

ConstructionDateText *	character	Date of construction when well was completed. May be year only.

WellDepthMeasure/MeasureValue *	character	Depth below land surface datum (LSD) to the bottom of the hole on completion of drilling. Measure value is given in the units stored in WellDepthMeasure/MeasureUnitCode.

WellDepthMeasure/MeasureUnitCode *	character	The code that represents the unit for measuring the item.

WellHoleDepthMeasure/MeasureValue *	character	Depth below land surface datum (LSD) to the bottom of the hole on completion of drilling. Measure value is given in the units stored in WellHoleDepthMeasure/MeasureUnitCode.

WellHoleDepthMeasure/MeasureUnitCode *	character	The code that represents the unit for measuring the item.

queryTime	POSIXct	Query time

* element is only in NWIS

Details

The readWQPsummary function has

See also

Examples

# \donttest{

site1 <- whatWQPsamples(siteid = "USGS-01594440")

type <- "Stream"

sites <- whatWQPsamples(countycode = "US:55:025", siteType = type)

lakeSites_samples <- whatWQPsamples(siteType = "Lake, Reservoir, Impoundment",
                                    countycode = "US:55:025")
# }
# \donttest{

type <- "Stream"

sites <- whatWQPmetrics(countycode = "US:55:025", siteType = type)
lakeSites_metrics <- whatWQPmetrics(siteType = "Lake, Reservoir, Impoundment",
                                    countycode = "US:55:025")
# }
# \donttest{

site1 <- whatWQPsites(siteid = "USGS-01594440")

type <- "Stream"
sites <- whatWQPsites(
  countycode = "US:55:025",
  characteristicName = "Phosphorus",
  siteType = type
)
# }